fix: clippy errors after update
As a result of the Rust toolchain upgrade from 1.65 to 1.66,
several clippy errors appeared (most of them being
clippy::unneccesary_cast). This commit fixes these errors.

(cherry picked from commit 0de96cee1e26e19ad741093068ae243aa9ed02ad)
(cherry picked from commit 8e2c2b9291962402f117a01fff39f1dd0966d771)

+++ src/devices/src/virtio/block/device.rs
@@ -96,7 +96,7 @@ impl DiskProperties {
             .map_err(Error::BackingFile)?;
         let disk_size = disk_image
             .seek(SeekFrom::End(0))
-            .map_err(Error::BackingFile)? as u64;
+            .map_err(Error::BackingFile)?;
 
         // We only support disk size, which uses the first two words of the configuration space.
         // If the image is not a multiple of the sector size, the tail bits are not exposed.
