+++ src/utils/src/time.rs
@@ -129,7 +129,7 @@ pub fn timestamp_cycles() -> u64 {
     #[cfg(target_arch = "x86_64")]
     // SAFETY: Safe because there's nothing that can go wrong with this call.
     unsafe {
-        std::arch::x86_64::_rdtsc() as u64
+        std::arch::x86_64::_rdtsc()
     }
     #[cfg(not(target_arch = "x86_64"))]
     {
